Modernization and Code Quality #2
Changes:
- command_prompt: Switch from "pyreadline" dependency to "prompt_toolkit"
- a2l_lex fix: Allow only exact matches of keywords for A2L types and sections
- lex tests: move lexer test cases
