+++ requirements.txt
@@ -1,5 +1,5 @@
 ply
 lxml
 loguru
-pyreadline
 alive-progress
+prompt-toolkit
