+++ tests/ast/test_ast_generator.py
@@ -22,7 +22,7 @@
 import os
 import tempfile
 import importlib.util
-from pathlib import Path
+from a2lparser import A2L_PACKAGE_DIR
 from a2lparser.a2l.ast.ast_generator import ASTGenerator
 
 
@@ -31,7 +31,7 @@ def test_unit_ast_generator():
     Attempts to create the python file containing the AST node classes.
     """
     with tempfile.TemporaryDirectory() as tempdir:
-        config_file = Path(__file__).parent / ".." / "a2lparser" / "configs" / "A2L_ASAM.cfg"
+        config_file = A2L_PACKAGE_DIR / "configs" / "A2L_ASAM.cfg"
         ast_python_file = os.path.join(tempdir, "test_a2l_ast.py")
         ast_generator = ASTGenerator(cfg_filename=str(config_file), out_filename=ast_python_file)
         ast_generator.generate(use_clean_names=True)
